refactor: onboarding api route
